refactor(blorktools): glb tool split
